refactor: 잘못된 자동차 이름 검증 로직을 cars와 car로 위임
