+++ src/main/java/racingcar/dto/CarsResponseDto.java
@@ -6,9 +6,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-//TODO
-// carResponseDtoList 이름 바꾸기 -> carResponseDtos? carsResponseDto?
-
 public class CarsResponseDto {
     private final List<CarResponseDto> carsResponseDto;
